created bottom menu component
