+++ src/main/java/com/solvd/demoapp/pages/common/PageBase.java
@@ -1,7 +1,7 @@
 package com.solvd.demoapp.pages.common;
 
+import com.solvd.demoapp.pages.components.bottommenu.BottomMenu;
 import com.zebrunner.carina.utils.ios.IOSUtils;
-import com.zebrunner.carina.webdriver.decorator.ExtendedWebElement;
 import com.zebrunner.carina.webdriver.gui.AbstractPage;
 import com.zebrunner.carina.webdriver.locator.ExtendedFindBy;
 import org.openqa.selenium.WebDriver;
@@ -11,46 +11,53 @@ import org.slf4j.LoggerFactory;
 public abstract class PageBase extends AbstractPage implements IOSUtils {
     private static final Logger LOGGER = LoggerFactory.getLogger(PageBase.class);
 
-    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"Catalog-tab-item\"`]")
-    private ExtendedWebElement catalogButton;
+//    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"Catalog-tab-item\"`]")
+//    private ExtendedWebElement catalogButton;
+//
+//    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"Cart-tab-item\"`]")
+//    private ExtendedWebElement cartButton;
+//
+//    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"More-tab-item\"`]")
+//    private ExtendedWebElement moreButton;
 
-    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"Cart-tab-item\"`]")
-    private ExtendedWebElement cartButton;
-
-    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeButton[`name == \"More-tab-item\"`]")
-    private ExtendedWebElement moreButton;
+    @ExtendedFindBy(iosClassChain = "**/XCUIElementTypeOther[`name == \"Catalog-screen\"`]/XCUIElementTypeOther[5]/XCUIElementTypeOther")
+    protected BottomMenu bottomMenu;
 
     public PageBase(WebDriver driver) {
         super(driver);
     }
 
-    public CatalogPageBase clickCatalogButton(){
-        LOGGER.info("clickCatalogButton()");
-        catalogButton.click();
-        return initPage(getDriver(), CatalogPageBase.class);
+    public BottomMenu getBottomMenu() {
+        return bottomMenu;
     }
 
-    public void clickCartButton(){
-        LOGGER.info("clickCartButton()");
-        cartButton.click();
-    }
-
-    public RightMenuPageBase clickMoreButton(){
-        LOGGER.info("clickMoreButton()");
-        moreButton.click();
-        return initPage(getDriver(), RightMenuPageBase.class);
-    }
-
-    public ExtendedWebElement getCatalogButton() {
-        return catalogButton;
-    }
-
-    public ExtendedWebElement getCartButton() {
-        return cartButton;
-    }
-
-    public ExtendedWebElement getMoreButton() {
-        return moreButton;
-    }
+    //    public CatalogPageBase clickCatalogButton(){
+//        LOGGER.info("clickCatalogButton()");
+//        catalogButton.click();
+//        return initPage(getDriver(), CatalogPageBase.class);
+//    }
+//
+//    public void clickCartButton(){
+//        LOGGER.info("clickCartButton()");
+//        cartButton.click();
+//    }
+//
+//    public RightMenuPageBase clickMoreButton(){
+//        LOGGER.info("clickMoreButton()");
+//        moreButton.click();
+//        return initPage(getDriver(), RightMenuPageBase.class);
+//    }
+//
+//    public ExtendedWebElement getCatalogButton() {
+//        return catalogButton;
+//    }
+//
+//    public ExtendedWebElement getCartButton() {
+//        return cartButton;
+//    }
+//
+//    public ExtendedWebElement getMoreButton() {
+//        return moreButton;
+//    }
 
 }
