+++ src/main/java/com/solvd/demoapp/pages/components/bottommenu/BottomMenu.java
@@ -1,0 +1,10 @@
+package com.solvd.demoapp.pages.components.bottommenu;
+
+import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.WebDriver;
+
+public class BottomMenu extends BottomMenuBase{
+    public BottomMenu(WebDriver driver, SearchContext searchContext) {
+        super(driver, searchContext);
+    }
+}
