+++ src/test/java/com/solvd/tests/demoapp/MyDemoAppTest.java
@@ -5,6 +5,7 @@ import com.solvd.demoapp.pages.common.CatalogPageBase;
 import com.solvd.demoapp.pages.common.LoginPageBase;
 import com.solvd.demoapp.pages.common.ProductPageBase;
 import com.solvd.demoapp.pages.common.RightMenuPageBase;
+import com.solvd.demoapp.pages.components.bottommenu.BottomMenu;
 import com.solvd.tests.AbstractTest;
 import com.zebrunner.carina.utils.R;
 import org.testng.annotations.Test;
@@ -18,10 +19,11 @@ public class MyDemoAppTest extends AbstractTest {
         //preconditions - user is logged - logging doesn't work on this app because of the broken keyboard feature!
         //it also doesn't work while buying!
         CatalogPageBase catalogPage = initPage(getDriver(), CatalogPageBase.class);
-        RightMenuPageBase rightMenuPage = catalogPage.clickMoreButton();
-        LoginPageBase loginPage = rightMenuPage.goToLoginPage();
-        loginPage.isPageOpened();
-        loginPage.logIn(R.TESTDATA.get("username-demo"), R.TESTDATA.get("password-demo"));
+        BottomMenu bottomMenu = catalogPage.getBottomMenu();
+        RightMenuPageBase rightMenuPageBase = bottomMenu.clickMoreButton();
+        LoginPageBase loginPage1 = rightMenuPageBase.goToLoginPage();
+        loginPage1.isPageOpened();
+        loginPage1.logIn(R.TESTDATA.get("username-demo"), R.TESTDATA.get("password-demo"));
     }
 
     @Test(testName = "#TC0002", description = "Validate that not logged user can add product to the cart and delete it")
